directive: move define tests
Let's split out the define tests and add a test for the previous commits
bug.
